exception handling

+++ src/main/java/com/vilaka/dscommerce/services/ProductService.java
@@ -3,9 +3,12 @@ package com.vilaka.dscommerce.services;
 import com.vilaka.dscommerce.dto.ProductDTO;
 import com.vilaka.dscommerce.entities.Product;
 import com.vilaka.dscommerce.repositories.ProductRepository;
-import com.vilaka.dscommerce.services.exceptions.ResourceNotFoundeption;
+import com.vilaka.dscommerce.services.exceptions.DataBaseException;
+import com.vilaka.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -22,7 +25,7 @@ public class ProductService {
     public ProductDTO findById(Long id){
 
         Optional<Product> result = repository.findById(id);
-        Product product = result.orElseThrow(()-> new ResourceNotFoundeption("Id não encontrado!"));
+        Product product = result.orElseThrow(()-> new ResourceNotFoundException("Recurso não encontrado!"));
 
         ProductDTO dto = new ProductDTO(product);
         return dto;
@@ -53,15 +56,29 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
     }
 
+    //@Transactional(Transactional.TxType.SUPPORTS)
+    //@Transactional(propagation = Propagation.SUPPORTS)
     @Transactional
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+        try{
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DataBaseException("Falha de integridade referêncial!");
+        }
+
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
